pyproject.toml
[project]
name = "tech-challenge-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "mlflow>=2.22.0",
    "numpy==1.26.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "uvicorn>=0.34.2",
    "yfinance>=0.2.61",
]
